requirements.txt
anaconda==0.0.1.1
appdirs==1.4.4
argon2-cffi==19.2.0
asgiref==3.5.2
autopep8==1.4.4
Babel==2.8.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
bootstrap4==0.1.0
braintree==4.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==37.0.4
defusedxml==0.6.0
distlib==0.3.4
dj-database-url==0.5.0
dj-stripe==2.3.0
Django==3.2
django-allauth==0.51.0
django-appconf==1.0.4
django-autoslug-iplweb==1.9.4.dev0
django-bootstrap4==1.1.1
django-braces==1.14.0
django-cities-light==3.6.0
django-ckeditor==6.4.2
django-countries==6.1.2
django-crispy-forms==1.9.0
django-daguerre==3.0.0
django-environ==0.9.0
django-generic-ratings==0.7
django-heroku==0.3.1
django-image-cropping==1.4.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-model-utils==4.0.0
django-modelcluster==5.0.1
django-node==4.0.2
django-paypal==1.0.0
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-responsive-images==1.2.0
django-social-share==2.3.0
django-star-ratings==0.9.1
django-storages==1.13.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
Faker==4.0.2
filelock==3.4.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
jsonfield==3.1.0
l18n==2018.5
mailchimp-marketing==3.0.75
mysqlclient @ file:///C:/Users/erick/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
nodejs==0.1.1
numpy==1.22.2
oauthlib==3.1.0
opencv-python==4.6.0.66
openpyxl==3.0.9
optional-django==0.1.0
pandas==1.4.1
pep8==1.7.1
phonenumbers==8.12.6
Pillow==7.0.0
platformdirs==2.4.0
portalocker==1.7.0
progressbar2==3.51.4
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pygame==2.1.2
PyJWT==2.4.0
python-dateutil==2.8.1
python-decouple==3.3
python-dhl==1.0.0.dev14
python-utils==2.4.0
python3-openid==3.2.0
pytz==2021.3
pywin32==227
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.2.4
stripe==2.41.0
suds-jurko==0.6
swapper==1.1.2.post1
text-unidecode==1.3
tzdata==2022.1
Unidecode==1.1.1
urllib3==1.25.9
virtualenv==20.10.0
wagtail==2.9
wagtail-lazyimages==0.1.8
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
xlrd==2.0.1
XlsxWriter==1.2.8
